adds docs, fixes a longitudinal python api issue

+++ src/feat.h
@@ -258,7 +258,7 @@ namespace FT{
 
             /// get longitudinal data from file s
             std::map<string, std::pair<vector<ArrayXd>, vector<ArrayXd>>> get_Z(string s, 
-                    long * idx, int idx_size);
+                    int * idx, int idx_size);
 
             /// destructor             
             ~Feat();
@@ -281,7 +281,7 @@ namespace FT{
 
             /// train a model, first loading longitudinal samples (Z) from file.
             void fit_with_z(double * X, int rowsX, int colsX, double * Y, int lenY, string s, 
-                            long * idx, int idx_size);
+                            int * idx, int idx_size);
            
             /// predict on unseen data.             
             VectorXd predict(MatrixXd& X,
@@ -302,11 +302,11 @@ namespace FT{
 	
             /// predict on unseen data, loading longitudinal samples (Z) from file.
             VectorXd predict_with_z(double * X, int rowsX,int colsX, 
-                                    string s, long * idx, int idx_size);
+                                    string s, int * idx, int idx_size);
 
             /// predict probabilities of each class.
             ArrayXXd predict_proba_with_z(double * X, int rowsX,int colsX, 
-                                    string s, long * idx, int idx_size);  
+                                    string s, int * idx, int idx_size);  
 
             /// predict on unseen data.             
             VectorXd predict(double * X, int rowsX, int colsX);      
@@ -321,7 +321,7 @@ namespace FT{
             
             /// train a model, first loading longitudinal samples (Z) from file.
             MatrixXd transform_with_z(double * X, int rowsX, int colsX, string s, 
-                                      long * idx, int idx_size);
+                                      int * idx, int idx_size);
             
             /// convenience function calls fit then predict.            
             VectorXd fit_predict(MatrixXd& X,
